move input improvements
add move input parsing to allow accurate game notation strings to be entered for input.

for example, a pawn attacking move can still be written as d4 (find a pawn to move to d4) but also as e.g. exd4 (move the pawn from 'e' to d4).
moves can be fed in 'as-is' from game notation, as only essential characters are parsed.

pieces to be moved will be disambiguated automatically if additional information was provided in the move notation.
for example Ngf6 means move the Knight at g to f6. if this information is not provided, a user choice prompt will still be shown as before.