checkmate method fixes
*  fix erroneous skip of location 0,0 rather than current king starting spot
*  in king single-move checking, fix so that king attacking moves are also considered. if the king can move one square in any direction from its current position in check, and not be in check afterwards, checkmate has not occurred. 
previously this was only checked if the square was completely free, but the same logic should apply if it contains an enemy piece. this handles potential situations where multiple attacking pieces attack on the same vector and an enemy non-attacking piece is in a surrounding square. for example a king may be in check with Rooks on either end of the row, with an escape route in the form of an unable to attack enemy Bishop directly in front of it. as the king can capture this Bishop and no longer be in check, checkmate would not occur.